Add favorite feature

+++ app/src/main/java/id/ac/ui/cs/mobileprogramming/ahmadsupriyanto/jahitcoid/database/FavoriteMovieDb.kt
@@ -19,7 +19,7 @@ data class FavoriteMovieDb (
     @ColumnInfo(name="original_language") val originalLanguage: String,
     @ColumnInfo(name="original_title") val originalTitle: String,
     @ColumnInfo(name="poster_path") val posterPath: String,
-    @ColumnInfo(name="popularity") val popularity: Float,
+    @ColumnInfo(name="popularity") val popularity: Double,
     @ColumnInfo(name="media_type") val mediaType: String,
     @ColumnInfo(name="vote_count") val voteCount: Int
 )
